Adds function to raise a dimensional physical quantity to an integer power.
Now we have two functions to raise (SI a b) to an integer power.
Old one, (^^) that was implemented earlier, requires dimensionles
quantity as a first argument but second argument is a normal
Integral number that could be set during run time. And new function,
that is introduced by this commit, does not require dimensionles
first operand but second operand should be known at compile time
to produce correct result type.