Moves calculation of the irreducible fractions to the later stage.
In the previous versions of the dimensional analysis procedure all
fractions were simlified to the irreducible form right after each
modification. That was an unnecessary action because our only concern
is to have irreducible fractions in the final result. This result could
be achieved by reducing fractions only during the late stage of
simplification process. Such modification reduces compile time and
required amount of memory.