Restores the old behaviour of test-compile suite.
Initially the test-compile suite has been using NoLink and HscNothing
options to only type-check provided test cases without actual compilation.
But that approach was not compatible with TemplateHaskell and test
program had to compile and link test cases to avoid runtime errors.
This problem was fixed in GHC 8.4.1 and compiling strategy for the
test cases could be revreted to its original version for up-to-date GHC.